backend/requirement.txt
fastapi
uvicorn
pydantic

numpy
sentence-transformers
torch
python-dotenv

